Create library for Google Measurement protocol
